Cleaner string suggestion
After including the word 'and' for translation, I think this is way more cleaner code and makes it easier for translation as it is one complete phrase. I think it's more coherent with the rest of your code also.
I'm sending all the changes needed in .pot and pt_PT verified language files.