Included all keywords possible in header and found strings
Strings "Add to Menu" and "Inspect Element" (#browser option) added.

Some strings got reordered, those from query mode/pre-path, etc, also in Show/Hide and Import/Export. Sorry for this. There is some little difference between your parsing and mine (Poedit) because the strings are all in the same lines. Still, the commitdiffs are very little.
I could reduce this differences editing the code manually, but I think that is not the way to go.