Added missing .po files for cs_CZ, pt_BR and ru_RU translations.
Method:
1) Opened existent .mo files in Vitraal, all texts appeared perfectly, can be compared with the some random strings that wordpress read from  the original files.
2) Just saved them again in Vitraal, apparently fixing it.
3) msgunfmt qtranslate-xx_XX.mo > qtranslate-xx_XX.po
4) Open in Poedit, everything looks ok.
5) Filled all header according current .pot catalog, update from this last .pot catalog to include not translated strings that are now shown as  empty.
6) Used the old existent strings "This function will not work correctly if you installed qTranslate on a blog with existing entries. In this  case you will need to take a look at \"Convert Database\" under \"Advanced Settings\"." and changed it to latest "This function will not work  correctly if you installed %s on a blog with existing entries. In this case you will need to take a look at \"%s\" under \"%s\"."
7) Save and compile .mo files.

Note: Some strings that do not correspond to existent ones are not so straight forward as "This function (...)". Theese strings are kept  automatically commented in the end of the .po file for further use. Some might be guessable, but I did not intend to start that work, I believe  that it can be done by those language translators. I just did one as example and because the %s vars where very obvious.