remove unsupported license key
The service handling the license key is not supported, since the host website is out of order.
The license form was recently introduced in the code, but never released officially.
All the code related to the license management is removed.
Empty code removed such as db version check.
Reverting activation hook back to version 3.4.7.