deprecate qtranxf_translate: explicit textdomains and colon translation
Some colon translations were missing (used at least for french).
The colon translation was confusing so it is changed for a more explicit template.

The internal translation calls were quite confusing.
It assumed some strings were translated in WordPress default textdomain but let
the possibility to override them.
Some strings expected in WordPress are actually missing.
Some internal calls were not even consistent with this method, bypassing qtranxf_translate.

To simplify we know use WordPress explicitly for those which are defined there.
For example: 'Settings' and 'Language'.
Those which are not defined require specific translation in the qtranslate domain.
As a consequence qtranxf_translate is removed for explicit calls and translations in the good domain.