prevent js cache issues in SCRIPT_DEBUG mode
When SCRIPT_DEBUG is true, the file timestamp is appended to the version
of the js script files to prevent issues with the client cache.
This is a quality-of-life feature for developers when working on the
non-minified versions (but it could be used for .min version as well).
The loading block is also moved out of the <script> block to better see
potential errors.