LSB wrap now made of a fixed class + optional extra classes
The previous options allowed to change the LSB wrap class(es) but now
with the Copy From behavior we need a fixed class for selection.
The extra classes should only be for styling and not the script behavior
so the wrap ul now always includes class "qtranxs-lang-switch-wrap".
The option becomes "additional LSB wrap classes".