deprecate lsb_style options wrap_class and active_class
The extra CSS classes for the LSB <ul>-wrap and active lang could lead to functional issues,
so they are removed from user options for simplification. These two options are deprecated:
- lsb_style_wrap_class,
- lsb_style_active_class.

Both values now depends directly on the LSB style for the built-in styles.

For custom CSS one should use "qtranxs-lang-switch-wrap" and "active" classes respectively,
which were the default values previously.
Note the "qtranxs-lang-switch-wrap" class is always included for the ul-wrap.

If you really need some specific customization please make a new feature request on github.
Meanwhile you should be able to use a built-in style.