refactor common.js, cleanup jQuery handler
This is a follow-up of the js cleanup for better consistency and readability:
- remove deprecated jQuery.ready(handler) since jQuery 3.0
- safe jQuery wrapper for standard internal usage with $()

The js qTranslateX class is now hidden from the global scope.
The related variable (hold in qTranslateConfig.qtx) is intanciated in this script instead of being out of control.
Here we assume common.js is enqueued before any other plugin script using it!
Other plugins should also follow the same common standards and not make qTranslate calls before DOM is loaded.

Note the global variable qTranslateConfig is unchanged.
It is still possible to access the qtx global variable through:
- qTranslateConfig.qtx
- qTranslateConfig.js.get_gtx()

This API is not very satisfying and it should be clarified with a more modern interface in the future.
If you are using this directly, please contact us (github) so we can plan these changes!