integrate plugin for woo-commerce 1.4
The integration plugin "woocommerce & qTranslate-X" is integrated from github.
The last release is 1.4 which is more recent than the 1.3 version on wordpress.org.

This plugin is now handled as an internal add-on in qTranslate-XT, seen as a module.

IMPORTANT: legacy integration plugin must not be activated!
Further changes will be made to ensure a better transition.

With this initial commit the only changes compared to the plugin are:
- reformat code style with standard wordpress PHP format.
- references to @since versions are removed.