Fix no redirect allowed on API REST (#621)
Prevent redirects on expected API REST calls from requests (with rest_get_url_prefix).

Adds rewrite rules for API REST with hidden lang.
This handles the special case when the default language is hidden in the options,
but asked in the request.

Note to make it work you have to flush your rewrite rules by saving the permalink
structures from the admin page!