Fix AMP url for themes like Penci Soledad
While I could not figure out the real cause (lack of time), this fixes
the localized URL string for modified URL links. It does so by looking
for a string like `/en/amp/en/` and replacing it as `/en/amp/`, so
both AMP and qtranslate-xt can be used in parallel.

This is tested with Penci Soledad AMP, but not with the official
AMP plugin from Wordpress/Google.