Dynamic module updates on plugin activation
The status of each module is now updated dynamically after activation/deactivation of the related plugins.
Previously the status of the modules was updated on every admin page, which was problematic if the user is not allowed to change admin settings.

For the activation of qTranslate plugin itself we check all the related plugins to update the module status.
For the other plugins events, one difficulty comes from a limitation of WP Core which prevents to use the usual "is_plugin_active" method.
On deactivation, the "active_plugins" option is updated only *after* all actions, though the hook should come after the effective deactivation (deactivated_plugin).
On activation, it is updated properly, unlike the deactivation which is inconsistent.
So we need to check dynamically each concerned plugin and handle the cases when the plugin activation state has just changed.