Cleanup obsolete admin versions notices
These notices were quite outdated and the versions comparison was cumbersome.
The version string was converted to int but this was only adding unnecessary complexity.
This feature may come back in the future but it should be made simpler with standard PHP "version_compare".
Keep in mind the options "qtranslate_version_previous" and "qtranslate_versions" may contain obsolete values.
They are only removed if the options are reset, otherwise kept (for potential qTranslate-X retro-compatibility).