Move date/time functions to qtranslate_date_time.php
Group all functions related to date/time from core and utils.
Not clear if they belong to a public API or only for private usage.
The 'public' versions seem to only be used internally in our hooks...

A few functions apparently not in use are commented out:
- qtranxf_formatCommentDateTime
- qtranxf_formatPostDateTime
- qtranxf_formatPostModifiedDateTime

Disclaimer: there might be some issues if those functions are used
directly.