Enable secure admin lang cookie (#467)
The "qtrans_admin_language" cookie is now created with the secure flag
if the option is checked.

It is now consistent with the "qtrans_front_language" cookie.

Note the option value still comes the "front settings" group, though
it also applies to the admin cookie here for simplicity.

This should be reworked, maybe with an automatic detection instead.