Update licence to GPLv2 or later
This is for clarification and harmonization for the WordPress community.
Note qTranslate-X had a mix of v2 and v3 but its owner cannot be reached.

This change should not be a problem since it remains GPL v2 *or later*.
If you ever redistribute this work you should also do it under the same license.

NB: "or later" matters, a downgrade to GPLv2-only would be more problematic,
since it would not allow to upgrade or include parts with GPLv3.