Remove WC get attributes hook (#612, #707)
This hook can't be used because of a internal limitation in WC with the
custom attributes only.
The option labels are read directly from DB and all the values that don't
match *exactly* the stored value are discarded. No other hook is called.
Note this problem doesn't concern the global attributes, treated as a
taxonomy.