Fix hreflang default URL no browser detect (#198, #819)
Only if the language is detected by the browser, enforce the
language shown to make the default URL deterministic for SEO.

Otherwise, allow the option "Hide URL language information
for default language".

Note the "x-default" item is not concerned by these changes,
this link always uses the default URL with no override.