Refactor module activation handler and options
Disambiguate the confusion between module state and admin options.

The module state depends in general on other plugin states.
The new admin options to manually disable or enable the modules are
just preferences for the module activation. They should be stored
regardless of the current state and can be deleted at any time.

Rename the option to `modules_ma_enabled` and make it admin option.
Refactor the module state update to separate the concepts.