Modules activation (#10)
* Refactor module activation handler and options

Disambiguate the confusion between module state and admin options.

The module state depends in general on other plugin states.
The new admin options to manually disable or enable the modules are
just preferences for the module activation. They should be stored
regardless of the current state and can be deleted at any time.

Rename the option to `modules_ma_enabled` and make it admin option.
Refactor the module state update to separate the concepts.

* Minor cleanup

* Fix missing checkbox, cleanup module info

* Refactor bool-array setting to `QTX_BOOLEAN_SET` (#1151)

The `bool_elements_array` setting was introduced in #1135 and #1137.
Re-using the QTX_ARRAY code with an extra argument makes it hard
to maintain. The QTX_ARRAY type is not suited for set of checkboxes.

Separate the logic by handling this type all separately.
The `QTX_BOOLEAN_SET` was legacy, defined but not used in the code.
Make the new type generic though only used by `ma_module_enabled`.
Remove the extra `bool_elements_array` argument.

* Use `QTX_BOOLEAN_SET` for module checkboxes

* Cleanup dead code

* Fix module init on activation, complete doc

* Rename `ma_enabled_modules` to `admin_enabled_modules`

Co-authored-by: HerrVigg <herrvigg@gmail.com>