Replace `strftime` with `IntlDateFormatter` for PHP8.1 (#1228)
`strftime` is deprecated in PHP8.1 and should not be used anymore.
Add a locale-formatted strftime using `IntlDateFormatter`.
This patch is a temporary workaround for immediate PHP8.1 support.

The `strftime` format should be abandonned completely in QTX so it
becomes legacy from now. Mark all strftime methods as deprecated for
PHPDoc but not for the PHP code yet, until a migration is done.

Attention: QTX uses an extended version of strftime to support extra
format tags such as `%q` that are non-standard in strftime.