Complete support of non-standard strftime (#1085)
Some formats like '%E' are not standard in strftime, but they have
been added in qTranslate to mimic some date formats.
Complete the conversion tables in the new `qxtranxf_intl_strftime`
to support those, for retro-compatibility. Anyway this function is
temporary, as strftime formats should not be used anymore.

Reorder the mappings in the original converter to better separate
non-standard pairs. Note that three of them are not supported in the
new function.