Deprecate date/strftime "override" options
The "override" date/time options are too confusing.
If a user ask for a specific format it should never be overriden.
The use case would only be for debug, but not for the end-user.
We might add new options to disable the translation or replace the
current language format with a default value, but that is out of
scope for the current "override" values.

Improve the CSS styles for deprecated options with two levels.