Fix broken ACF init sequence for options LSB (#1233)
A regression was introduced in 3.12.1, because the ACF admin was
setting its `qtranslate_admin_config` hook too late. That filter
must be set before action  `qtranslate_init_language` but it can't
be done with `plugins_loaded`.
This patch restores this hook earlier when the module is loaded,
by calling `qtranxf_acf_init` to do the full init at that time.
The loading sequence will be revisited for all modules.