Keep hidden elements with detached field (#1253)
Though the `form` attribute is quite unusal, it becomes harder to
find the hidden elements if those are created under the form.
Better keep them together with the detached input field.

The `qtranslate-edit-language` is an exception as it's unique for
the whole form and not linked to other elements.

Simplify the code by re-using `qtranxf_ce`. Extend it by sending
the `beforeNode` instead of a boolean handling only the firstNode.