Generalize ACF config and simplify options
The ACF options are difficult to grasp:
- it should not be required to enter the name of the pages where
  the LSB are shown, these should be defined by the module.
- the standard ACF input and text fields should always be supported.
  We should deprecate the ACF-QTX extended fields.

Moreover the `shim` element used to set the LSB anchors is too fuzzy.
It may appear several times with same id in the DOM.
Also, it's unclear how the display hooks are handled, as they should
not be dependent on the options.

Refactor the ACF configuration to list clearly the supported pages,
anchors and display fields.

Remove the options for standard support (always true) and the list
of pages for LSB (the config should be updated).