Refactor admin CSS and LSB files
The CSS files for LSB should not be processed and embedded every time.
This complicates the code and it's not efficient for the cache.

Take the LSB CSS out of the embedded CSS and load them as CSS files.

Rename the CSS files in lower case separated by hyphens.

Rename the main `qtranslate_configuration.css` as `admin.css`.