Refactor admin CSS and LSB files (#1294)
The CSS files for LSB should not be processed and embedded every time. This complicates the code and it's not efficient for the cache.
Load the LSB CSS as files as it's not dependent on embedded code (for user color).
Enable localization of LSB options.
Rename the CSS files in lower case separated by hyphens.
Rename the main `qtranslate_configuration.css` as `admin.css`.
Tidy up highlight CSS code.