Add new ACF settings to select extended types
Eventually the extended types may be removed completely.
This new feature allows to choose which extended types can be picked.
The default value becomes false for all extended types.

This only impacts the list of types available in the group edit.
The extended types are still loaded and translated as usual.
It is up to the admin to clean up pending types.