Deprecate `wp_translator` filter (#1304)
This legacy filter should be named with a `qtranslate` prefix.
It is not well designed. The intention was to allow third-party
integration of a custom Translator class using the i18n interface
but there is no practical way to change it.

Deprecate this filter, to be removed in next major release.
If there is a real need it should be done differently.