Revert PHP types in `qtranxf_checkCanonical` (#1314)
The main purpose is to filter `redirect_canonical` (WP hook).
The WP documentation is not precise regarding the types so we have
to relax the types check. Check first if the redirect URL is false.

The second parameter is the requested URL but it is ignored.
See #1320, wrong usage with Yoast. The second parameter has a
different meaning with Yoast. Removing the types will work but the
Yoast hook should be rewritten with a different function.

Rewrite the function to handle the default language in convertURL.