Remove legacy `pg` argument and props in JS QTX
The `pg` parameter to the QTranslateX pseudo-class has become obsolete.
It was used in early versions of qTranslate-X for the first versions of
LSB to carry these props:
- addContentHooks
- langSwitchWrapAnchor
- onTabSwitch

These have been replaced and extended long ago by the JSON config passed
via `qTranslateConfig.page_config`. There is no PHP code setting them
and they are not documented parameters so let's get rid of this.

See 2.9.8.0 pre-release for reference.