Replace JS `qTranslateX` object with `hooks` module (#1458)
 The `qTranslateX` pseudo-class does not make much sense as it's
only instantiated once. Change it to a proper module API now that
we can bundle it with webpack/node. 

The new API becomes;
- `qTranx`: main module for generic methods (to be extended)
- `qTranx.hooks`: new sub-module for ML content and LSB,
providing the same API as the `qTranslateX` object.

Deprecate `qTranslateConfig.js.get_qtx`.
The legacy `qtx` object is just holding the ML hooks and should
no longer be used. This will be removed in next major release.
To avoid the deprecation message simply use `qTranx.hooks`
as the former `qtx` object (it was the main API).

For developers this change should have no impact as the old
interface is still preserved with a few differences:
- An `init` function is now called explicitly by the internal loaders. 
- The `hooks` module is now available immediately.
- However it should not be used until the `qtxLoadAdmin` event is
fired similar to what was expected with the `get_qtx` method.