Handle QTX custom fields internally
Make the custom fields config private (WIP).
Deprecate `addCustomContentHooks` QTX legacy API.
It was triggered for `post` page but this API creates complications.
Custom fields are confusing as they are globals for IDs but not for
class selectors, though they are documented to be triggered from
`post` page only. In fact they should be part of the `post` page
config, but they are not linked to a form.
Move the selector logic from JS post page to PHP footer config and
generalize the custom hook creation to the main JS hooks init.