Generalize admin notices for deprecated settings (#1470)
Notices for deprecated settings were spread over many places and
creating integration issues in WP. They must only be displayed
on `admin_notices` event.

Regroup all deprecated settings in one place with a generic handler
so it can be easily extended.

Fix a bug with the JS notice handler to send proper ajax queries
when the cross button is clicked (not assuming a div button exists).